chore: move utils/wildcard in ext (#8772)
