+++ ext/wildcard/utils.go
