+++ ext/wildcard/utils_test.go
