+++ pkg/utils/kube/wildcard.go
@@ -1,7 +1,7 @@
 package kube
 
 import (
-	"github.com/kyverno/kyverno/pkg/utils/wildcard"
+	"github.com/kyverno/kyverno/ext/wildcard"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
