+++ pkg/utils/match/name.go
@@ -1,7 +1,7 @@
 package match
 
 import (
-	"github.com/kyverno/kyverno/pkg/utils/wildcard"
+	"github.com/kyverno/kyverno/ext/wildcard"
 )
 
 func CheckName(expected, actual string) bool {
